requirements.txt
backcall==0.2.0
Bottleneck==1.3.2
certifi==2021.10.8
cycler==0.10.0
decorator==5.1.0
fonttools==4.25.0
ipdb==0.13.9
ipython==7.29.0
jedi==0.18.1
kiwisolver==1.3.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
numexpr==2.7.3
numpy==1.21.2
olefile==0.46
opencv-python==4.5.4.60
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.2.4
prompt-toolkit==3.0.22
ptyprocess==0.7.0
Pygments==2.10.0
pyparsing==3.0.4
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
scipy==1.7.1
setuptools==58.0.4
six==1.16.0
toml==0.10.2
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.10.0.2
wcwidth==0.2.5
wheel==0.37.0
